+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/adapter/StudyPeopleViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.created.team201.databinding.ItemStudyDetailStudyPeopleBinding
-import com.created.team201.presentation.studyList.uiModel.StudyParticipant
+import com.created.team201.presentation.studyDetail.model.StudyParticipant
 
 class StudyPeopleViewHolder(val binding: ItemStudyDetailStudyPeopleBinding) :
     RecyclerView.ViewHolder(binding.root) {
