+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/model/StudyDetailUIModel.kt
@@ -1,6 +1,6 @@
 package com.created.team201.presentation.studyList.uiModel
 
-data class StudyUIModel(
+data class StudyDetailUIModel(
     val title: String,
     val introduction: String,
     val peopleCount: Int,
