+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/model/StudyParticipant.kt
@@ -1,9 +1,9 @@
-package com.created.team201.presentation.studyList.uiModel
+package com.created.team201.presentation.studyDetail.model
 
 data class StudyParticipant(
     val isMaster: Boolean,
     val profileImageUrl: String,
     val name: String,
     val successRate: Int,
-    val tier: String,
+    val tier: Int,
 )
